+++ src/Stack/PackageLocation.hs
@@ -9,11 +9,8 @@
 -- | Deal with downloading, cloning, or whatever else is necessary for
 -- getting a 'PackageLocation' into something Stack can work with.
 module Stack.PackageLocation
-  ( resolveSinglePackageLocation
-  , resolveMultiPackageLocation
-  , parseSingleCabalFile
+  ( parseSingleCabalFile
   , parseSingleCabalFileIndex
-  , parseMultiCabalFiles
   , parseMultiCabalFilesIndex
   ) where
 
@@ -41,196 +38,21 @@ import Stack.Types.PackageIdentifier
 import qualified System.Directory as Dir
 import RIO.Process
 
--- | Same as 'resolveMultiPackageLocation', but works on a
--- 'SinglePackageLocation'.
-resolveSinglePackageLocation
-    :: HasConfig env
-    => Path Abs Dir -- ^ project root
-    -> PackageLocation FilePath
-    -> RIO env (Path Abs Dir)
-resolveSinglePackageLocation projRoot (PLFilePath fp) = resolveDir projRoot fp
-resolveSinglePackageLocation projRoot (PLArchive (Archive url subdir msha)) = do
-    workDir <- view workDirL
+flattenPackageLocation :: Traversable t => t Subdirs -> [t FilePath]
+flattenPackageLocation =
+    traverse go
+  where
+    go :: Subdirs -> [FilePath]
+    go DefaultSubdirs = [""]
+    go (ExplicitSubdirs subs) = map go' subs
 
-        -- TODO: dedupe with code for snapshot hash?
-    let name = T.unpack $ decodeUtf8 $ S.take 12 $ B64URL.encode $ Mem.convert $ hashWith SHA256 $ encodeUtf8 url
-        root = projRoot </> workDir </> $(mkRelDir "downloaded")
-        fileExtension' = ".http-archive"
+    go' :: FilePath -> FilePath
+    go' = T.unpack
+        . T.intercalate "/"
+        . filter (\t -> not (T.null t) && t /= ".")
+        . T.split (== '/')
+        . T.pack
 
-    fileRel <- parseRelFile $ name ++ fileExtension'
-    dirRel <- parseRelDir name
-    dirRelTmp <- parseRelDir $ name ++ ".tmp"
-    let fileDownload = root </> fileRel
-        dir = root </> dirRel
-
-    exists <- doesDirExist dir
-    unless exists $ do
-        liftIO $ ignoringAbsence (removeDirRecur dir)
-
-        let dirTmp = root </> dirRelTmp
-        liftIO $ ignoringAbsence (removeDirRecur dirTmp)
-
-        urlExists <- liftIO $ Dir.doesFileExist $ T.unpack url
-        file <-
-          if urlExists
-            then do
-              file <- liftIO $ Dir.canonicalizePath (T.unpack url) >>= parseAbsFile
-              case msha of
-                Nothing -> return ()
-                Just sha -> do
-                  actualSha <- mkStaticSHA256FromFile $ toFilePath file
-                  when (sha /= actualSha) $ error $ concat
-                    [ "Invalid SHA256 found for local archive "
-                    , show file
-                    , "\nExpected: "
-                    , T.unpack $ staticSHA256ToText sha
-                    , "\nActual:   "
-                    , T.unpack $ staticSHA256ToText actualSha
-                    ]
-              return file
-            else do
-              req <- parseUrlThrow $ T.unpack url
-              let dreq = DownloadRequest
-                    { drRequest = req
-                    , drHashChecks =
-                        case msha of
-                          Nothing -> []
-                          Just sha ->
-                            [HashCheck
-                              { hashCheckAlgorithm = SHA256
-                              , hashCheckHexDigest = CheckHexDigestByteString $ staticSHA256ToBase16 sha
-                              }]
-                    , drLengthCheck = Nothing -- TODO add length info?
-                    , drRetryPolicy = drRetryPolicyDefault
-                    }
-              _ <- verifiedDownload dreq fileDownload (const $ return ())
-              return fileDownload
-
-        let fp = toFilePath file
-
-        withLazyFile fp $ \lbs -> do
-          -- Share a single file read among all of the different
-          -- parsing attempts. We're not worried about unbounded
-          -- memory usage, as we will detect almost immediately if
-          -- this is the wrong type of file.
-
-          let tryTargz = do
-                logDebug $ "Trying to ungzip/untar " <> fromString fp
-                let entries = Tar.read $ GZip.decompress lbs
-                liftIO $ Tar.unpack (toFilePath dirTmp) entries
-              tryZip = do
-                logDebug $ "Trying to unzip " <> fromString fp
-                let archive = Zip.toArchive lbs
-                liftIO $  Zip.extractFilesFromArchive [Zip.OptDestination
-                                                       (toFilePath dirTmp)] archive
-              tryTar = do
-                logDebug $ "Trying to untar (no ungzip) " <> fromString fp
-                let entries = Tar.read lbs
-                liftIO $ Tar.unpack (toFilePath dirTmp) entries
-              err = throwM $ UnableToExtractArchive url file
-
-              catchAnyLog goodpath handler =
-                  catchAny goodpath $ \e -> do
-                      logDebug $ "Got exception: " <> displayShow e
-                      handler
-
-          tryTargz `catchAnyLog` tryZip `catchAnyLog` tryTar `catchAnyLog` err
-        renameDir dirTmp dir
-
-    x <- listDir dir
-    case x of
-        ([dir'], []) -> resolveDir dir' subdir
-        (dirs, files) -> liftIO $ do
-            ignoringAbsence (removeFile fileDownload)
-            ignoringAbsence (removeDirRecur dir)
-            throwIO $ UnexpectedArchiveContents dirs files
-resolveSinglePackageLocation projRoot (PLRepo (Repo url commit repoType' subdir)) =
-    cloneRepo projRoot url commit repoType' >>= flip resolveDir subdir
-
--- | Resolve a PackageLocation into a path, downloading and cloning as
--- necessary.
---
--- Returns the updated PackageLocation value with just a single subdir
--- (if relevant).
-resolveMultiPackageLocation
-    :: HasConfig env
-    => Path Abs Dir -- ^ project root
-    -> PackageLocation Subdirs
-    -> RIO env [(Path Abs Dir, PackageLocation FilePath)]
-resolveMultiPackageLocation y (PLFilePath fp) = do
-  dir <- resolveSinglePackageLocation y (PLFilePath fp)
-  return [(dir, PLFilePath fp)]
-resolveMultiPackageLocation y (PLArchive (Archive url subdirs msha)) = do
-  dir <- resolveSinglePackageLocation y (PLArchive (Archive url "." msha))
-  let subdirs' =
-        case subdirs of
-          DefaultSubdirs -> ["."]
-          ExplicitSubdirs subs -> subs
-  forM subdirs' $ \subdir -> do
-    dir' <- resolveDir dir subdir
-    return (dir', PLArchive (Archive url subdir msha))
-resolveMultiPackageLocation projRoot (PLRepo (Repo url commit repoType' subdirs)) = do
-  dir <- cloneRepo projRoot url commit repoType'
-
-  let subdirs' =
-        case subdirs of
-          DefaultSubdirs -> ["."]
-          ExplicitSubdirs subs -> subs
-  forM subdirs' $ \subdir -> do
-    dir' <- resolveDir dir subdir
-    return (dir', PLRepo $ Repo url commit repoType' subdir)
-
-cloneRepo
-    :: HasConfig env
-    => Path Abs Dir -- ^ project root
-    -> Text -- ^ URL
-    -> Text -- ^ commit
-    -> RepoType
-    -> RIO env (Path Abs Dir)
-cloneRepo projRoot url commit repoType' = do
-    workDir <- view workDirL
-    let nameBeforeHashing = case repoType' of
-            RepoGit -> T.unwords [url, commit]
-            RepoHg -> T.unwords [url, commit, "hg"]
-        -- TODO: dedupe with code for snapshot hash?
-        name = T.unpack $ decodeUtf8 $ S.take 12 $ B64URL.encode $ Mem.convert $ hashWith SHA256 $ encodeUtf8 nameBeforeHashing
-        root = projRoot </> workDir </> $(mkRelDir "downloaded")
-
-    dirRel <- parseRelDir name
-    let dir = root </> dirRel
-
-    exists <- doesDirExist dir
-    unless exists $ do
-        liftIO $ ignoringAbsence (removeDirRecur dir)
-
-        let cloneAndExtract commandName cloneArgs resetCommand =
-              withWorkingDir (toFilePath root) $ do
-                ensureDir root
-                logInfo $ "Cloning " <> display commit <> " from " <> display url
-                proc commandName
-                       ("clone" :
-                        cloneArgs ++
-                        [ T.unpack url
-                        , toFilePathNoTrailingSep dir
-                        ]) runProcess_
-                created <- doesDirExist dir
-                unless created $ throwM $ FailedToCloneRepo commandName
-                withWorkingDir (toFilePath dir) $ readProcessNull commandName
-                    (resetCommand ++ [T.unpack commit, "--"])
-                    `catchAny` \case
-                        ex -> do
-                            logInfo $
-                              "Please ensure that commit " <>
-                              display commit <>
-                              " exists within " <>
-                              display url
-                            throwM ex
-
-        case repoType' of
-            RepoGit -> cloneAndExtract "git" ["--recursive"] ["--git-dir=.git", "reset", "--hard"]
-            RepoHg  -> cloneAndExtract "hg"  []              ["--repository", ".", "update", "-C"]
-
-    return dir
 
 -- | Parse the cabal files present in the given
 -- 'PackageLocationIndex FilePath'.
@@ -244,39 +66,21 @@ parseSingleCabalFileIndex
 -- index tarball) and correctness (get the cabal file from the index,
 -- not the package tarball itself, yay Hackage revisions).
 parseSingleCabalFileIndex _ (PLIndex pir) = readPackageUnresolvedIndex pir
-parseSingleCabalFileIndex root (PLOther loc) = lpvGPD <$> parseSingleCabalFile root False loc
+parseSingleCabalFileIndex root (PLOther (PLFilePath fp)) = do
+  dir <- resolveDir root fp
+  lpvGPD <$> parseSingleCabalFile False dir
 
 parseSingleCabalFile
   :: forall env. HasConfig env
-  => Path Abs Dir -- ^ project root, used for checking out necessary files
-  -> Bool -- ^ print warnings?
-  -> PackageLocation FilePath
+  => Bool -- ^ print warnings?
+  -> Path Abs Dir
   -> RIO env LocalPackageView
-parseSingleCabalFile root printWarnings loc = do
-  dir <- resolveSinglePackageLocation root loc
+parseSingleCabalFile printWarnings dir = do
   (gpd, cabalfp) <- readPackageUnresolvedDir dir printWarnings
   return LocalPackageView
     { lpvCabalFP = cabalfp
     , lpvGPD = gpd
-    , lpvLoc = loc
     }
-
--- | Load and parse cabal files into 'GenericPackageDescription's
-parseMultiCabalFiles
-  :: forall env. HasConfig env
-  => Path Abs Dir -- ^ project root, used for checking out necessary files
-  -> Bool -- ^ print warnings?
-  -> PackageLocation Subdirs
-  -> RIO env [LocalPackageView]
-parseMultiCabalFiles root printWarnings loc0 =
-  resolveMultiPackageLocation root loc0 >>=
-  mapM (\(dir, loc1) -> do
-    (gpd, cabalfp) <- readPackageUnresolvedDir dir printWarnings
-    return LocalPackageView
-      { lpvCabalFP = cabalfp
-      , lpvGPD = gpd
-      , lpvLoc = loc1
-      })
 
 -- | 'parseMultiCabalFiles' but supports 'PLIndex'
 parseMultiCabalFilesIndex
@@ -284,9 +88,5 @@ parseMultiCabalFilesIndex
   => Path Abs Dir -- ^ project root, used for checking out necessary files
   -> PackageLocationIndex Subdirs
   -> RIO env [(GenericPackageDescription, PackageLocationIndex FilePath)]
-parseMultiCabalFilesIndex _root (PLIndex pir) =
-  pure . (, PLIndex pir) <$>
-  readPackageUnresolvedIndex pir
-parseMultiCabalFilesIndex root (PLOther loc0) =
-  map (\lpv -> (lpvGPD lpv, PLOther $ lpvLoc lpv)) <$>
-  parseMultiCabalFiles root False loc0
+parseMultiCabalFilesIndex root pl0 = for (flattenPackageLocation pl0) $ \pl ->
+  (, pl) <$> parseSingleCabalFileIndex root pl